requirements.txt
binance_futures
numpy==1.20.3
tulipy==0.4.0
websocket_client==1.2.3
